+++ src/components/Button/index.ts
@@ -1,4 +1,6 @@
 import Button from './Button';
 import ButtonMode from './ButtonMode';
+import ButtonText from './ButtonText';
+import ButtonOutline from './ButtonOutline';
 
-export {Button, ButtonMode};
+export {Button, ButtonMode, ButtonText, ButtonOutline};
