--- src/components/ButtonOutline/index.ts
@@ -1,3 +1,0 @@
-import ButtonOutline from './ButtonOutline';
-
-export default ButtonOutline;
