--- src/components/ButtonText/index.ts
@@ -1,3 +1,0 @@
-import ButtonText from './ButtonText';
-
-export default ButtonText;
